project: rename project to hurd
