+++ hurd/errors.py
