+++ hurd/platform.py
